Refactor server side functions to seperate mongoDB connect from reading and writing to the database
