+++ Server/server.js
@@ -4,7 +4,9 @@ const axios = require('axios');
 const fs = require('fs');
 const APILinks = JSON.parse(fs.readFileSync('./data/APIData.json'));
 const PUBLICATIONS = require('./publications.js');
-const addDataDB = require('./mongodb.js');
+const updateDB = require('./updateDB.js');
+const connectDB = require('./connectDB.js');
+const readDB = require('./readDB.js');
 const app = express();
 
 app.use(express.json());
@@ -13,7 +15,7 @@ app.use(cors());
 // This can stay the same?
 app.get('/', (req, res) => {
     const { publication } = req.query;
-    const data = fetchData(publication);
+    const data = readDB(publication);
     res.json(data);
 })
 
@@ -22,34 +24,34 @@ app.get('/', (req, res) => {
 // 2. Query through DB and return all posts that match the query request
 // 2.5. Organize posts based on posted_time value
 // 3. Include option to query for ALL publications simultaneously
-const fetchData = (publication) => {
-    switch(publication) {
-        case(PUBLICATIONS.GLOBAL_TORONTO): {
-            const data = fs.readFileSync('./data/global.json');
-            return JSON.parse(data);
-        }
-        case(PUBLICATIONS.GLOBE_AND_MAIL): {
-            const data = fs.readFileSync('./data/globe.json');
-            return JSON.parse(data);
-        }
-        case(PUBLICATIONS.TORONTO_LIFE): {
-            const data = fs.readFileSync('./data/tlife.json');
-            return JSON.parse(data);
-        }
-        case(PUBLICATIONS.WALRUS): {
-            const data = fs.readFileSync('./data/walrus.json');
-            return JSON.parse(data);
-        }
-        case(PUBLICATIONS.CANADALAND): {
-            const data = fs.readFileSync('./data/cland.json');
-            return JSON.parse(data);
-        }
-        case(PUBLICATIONS.NOW_TORONTO): {
-            const data = fs.readFileSync('./data/now.json');
-            return JSON.parse(data);
-        }
-    }
-}
+// const fetchData = (publication) => {
+//     switch(publication) {
+//         case(PUBLICATIONS.GLOBAL_TORONTO): {
+//             const data = fs.readFileSync('./data/global.json');
+//             return JSON.parse(data);
+//         }
+//         case(PUBLICATIONS.GLOBE_AND_MAIL): {
+//             const data = fs.readFileSync('./data/globe.json');
+//             return JSON.parse(data);
+//         }
+//         case(PUBLICATIONS.TORONTO_LIFE): {
+//             const data = fs.readFileSync('./data/tlife.json');
+//             return JSON.parse(data);
+//         }
+//         case(PUBLICATIONS.WALRUS): {
+//             const data = fs.readFileSync('./data/walrus.json');
+//             return JSON.parse(data);
+//         }
+//         case(PUBLICATIONS.CANADALAND): {
+//             const data = fs.readFileSync('./data/cland.json');
+//             return JSON.parse(data);
+//         }
+//         case(PUBLICATIONS.NOW_TORONTO): {
+//             const data = fs.readFileSync('./data/now.json');
+//             return JSON.parse(data);
+//         }
+//     }
+// }
 
 const pullDataAndSave = async () => {
 
@@ -68,7 +70,7 @@ const pullDataAndSave = async () => {
 
 const postObjects = await Promise.all(pullPromises)
 
-await addDataDB(postObjects.flat(1))
+await updateDB(postObjects.flat(1))
 }
 pullDataAndSave();
 
@@ -79,4 +81,5 @@ const autoAPICall = setInterval(function() {
 
 app.listen(8080, () => {
     console.log('Listening on port: 8080')
+    connectDB();
 })
